ingester/requirements.txt
azure-functions==1.5.0
pyodbc==4.0.30
python-dotenv==0.15.0
